Drop ios test
