+++ src/enamlnative/android/android_view_group.py
@@ -9,10 +9,9 @@ Created on May 20, 2017
 
 @author: jrm
 """
-from atom.api import Typed, Subclass, set_default
+from atom.api import Typed, set_default
 
 from enamlnative.widgets.view_group import ProxyViewGroup
-from enamlnative.widgets.view import coerce_size
 
 from .android_view import AndroidView, View, LayoutParams
 from .bridge import JavaBridgeObject, JavaMethod
@@ -52,7 +51,7 @@ class AndroidViewGroup(AndroidView, ProxyViewGroup):
     widget = Typed(ViewGroup)
 
     #: Layout type
-    layout_param_type = Subclass(LayoutParams, default=MarginLayoutParams)
+    layout_param_type = set_default(MarginLayoutParams)
 
     #: Default layout params
     default_layout = set_default({
@@ -84,6 +83,10 @@ class AndroidViewGroup(AndroidView, ProxyViewGroup):
                 else:
                     widget.addView(child_widget, i)
                 i += 1
+
+        # Force layout using the default params
+        if not self.layout_params:
+            self.set_layout({})
 
     def child_added(self, child):
         """ Handle the child added event from the declaration.
@@ -123,90 +126,6 @@ class AndroidViewGroup(AndroidView, ProxyViewGroup):
         super(AndroidViewGroup, self).child_removed(child)
         if child.widget is not None:
             self.widget.removeView(child.widget)
-
-    def create_layout_params(self, child, layout):
-        """ Create the LayoutParams for a child with it's requested
-        layout parameters. Subclasses should override this as needed
-        to handle layout specific needs.
-        
-        Parameters
-        ----------
-        child: AndroidView
-            A view to create layout params for.
-        layout: Dict
-            A dict of layout parameters to use to create the layout.
-             
-        Returns
-        -------
-        layout_params: LayoutParams
-            A LayoutParams bridge object with the requested layout options.
-        
-        """
-        dp = self.dp
-        w, h = (coerce_size(layout.get('width', 'wrap_content')),
-                coerce_size(layout.get('height', 'wrap_content')))
-        w = w if w < 0 else int(w * dp)
-        h = h if h < 0 else int(h * dp)
-        layout_params = self.layout_param_type(w, h)
-
-        if layout.get('margin'):
-            l, t, r, b = layout['margin']
-            layout_params.setMargins(int(l*dp), int(t*dp),
-                                     int(r*dp), int(b*dp))
-        return layout_params
-
-    def apply_layout(self, child, layout):
-        """ Apply a layout to a child. This sets the layout_params
-        of the child which is later used during the `init_layout` pass.
-        Subclasses should override this as needed to handle layout specific
-        needs of the ViewGroup.
-        
-        Parameters
-        ----------
-        child: AndroidView
-            A view to create layout params for.
-        layout: Dict
-            A dict of layout parameters to use to create the layout.
-        
-        """
-        layout_params = child.layout_params
-        if not layout_params:
-            layout_params = self.create_layout_params(child, layout)
-        w = child.widget
-        if w:
-            dp = self.dp
-            # padding
-            if 'padding' in layout:
-                l, t, r, b = layout['padding']
-                w.setPadding(int(l*dp), int(t*dp),
-                             int(r*dp), int(b*dp))
-
-            # left, top, right, bottom
-            if 'left' in layout:
-                w.setLeft(int(layout['left']*dp))
-            if 'top' in layout:
-                w.setTop(int(layout['top']*dp))
-            if 'right' in layout:
-                w.setRight(int(layout['right']*dp))
-            if 'bottom' in layout:
-                w.setBottom(int(layout['bottom']*dp))
-
-            # x, y, z
-            if 'x' in layout:
-                w.setX(layout['x']*dp)
-            if 'y' in layout:
-                w.setY(layout['y']*dp)
-            if 'z' in layout:
-                w.setZ(layout['z']*dp)
-
-            # set min width and height
-            # maximum is not supported by AndroidViews (without flexbox)
-            if 'min_height' in layout:
-                w.setMinimumHeight(int(layout['min_height']*dp))
-            if 'min_width' in layout:
-                w.setMinimumWidth(int(layout['min_width']*dp))
-
-        child.layout_params = layout_params
 
     def set_transition(self, transition):
         t = LayoutTransition() if transition == 'default' else None
