+++ tests/test_examples.py
@@ -50,7 +50,7 @@ from utils import load
     (["android"], 'toast.enaml'),
     (["android"], 'view_pager.enaml'),
     (["android"], 'webview.enaml'),
-    (["ios"], '../ios/demo/Python/view.enaml'),
+    #(["ios"], '../ios/demo/Python/view.enaml'),
 ])
 def test_examples(platforms, path):
     #: Load
